Rearrange JVM code to keep deprecated stuff separate
